Refacot NodeElement
