+++ src/tree.jquery.ts
@@ -11,7 +11,9 @@ import SelectNodeHandler from "./selectNodeHandler";
 import SimpleWidget from "./simple.widget";
 import { Node, getPosition } from "./node";
 import { isFunction } from "./util";
-import { FolderElement, NodeElement, OnFinishOpenNode } from "./nodeElement";
+import NodeElement from "./nodeElement";
+import FolderElement, { OnFinishOpenNode } from "./nodeElement/folderElement";
+
 import { JQTreeOptions } from "./jqtreeOptions";
 
 interface ClickTarget {
@@ -101,7 +103,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
 
     public selectNode(
         node: Node | null,
-        optionsParam?: SelectNodeOptions
+        optionsParam?: SelectNodeOptions,
     ): JQuery {
         this.doSelectNode(node, optionsParam);
         return this.element;
@@ -137,21 +139,21 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
     public loadDataFromUrl(
         param1: string | null | Node,
         param2?: Node | null | HandleFinishedLoading,
-        param3?: HandleFinishedLoading
+        param3?: HandleFinishedLoading,
     ): JQuery {
         if (typeof param1 === "string") {
             // first parameter is url
             this.doLoadDataFromUrl(
                 param1,
                 param2 as Node | null,
-                param3 ?? null
+                param3 ?? null,
             );
         } else {
             // first parameter is not url
             this.doLoadDataFromUrl(
                 null,
                 param1,
-                param2 as HandleFinishedLoading | null
+                param2 as HandleFinishedLoading | null,
             );
         }
 
@@ -185,7 +187,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
     }
 
     public getNodeByHtmlElement(
-        element: HTMLElement | JQuery<HTMLElement>
+        element: HTMLElement | JQuery<HTMLElement>,
     ): Node | null {
         return this.getNode(jQuery(element));
     }
@@ -197,7 +199,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
     public openNode(
         node: Node,
         param1?: boolean | OnFinishOpenNode,
-        param2?: OnFinishOpenNode
+        param2?: OnFinishOpenNode,
     ): JQuery {
         if (!node) {
             throw Error(NODE_PARAM_IS_EMPTY);
@@ -238,7 +240,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
         if (node.isFolder() || node.isEmptyFolder) {
             new FolderElement(node, this).close(
                 slide,
-                this.options.animationSpeed
+                this.options.animationSpeed,
             );
 
             this.saveState();
@@ -258,7 +260,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
 
     public addNodeAfter(
         newNodeInfo: NodeData,
-        existingNode: Node
+        existingNode: Node,
     ): Node | null {
         const newNode = existingNode.addAfter(newNodeInfo);
 
@@ -271,7 +273,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
 
     public addNodeBefore(
         newNodeInfo: NodeData,
-        existingNode: Node
+        existingNode: Node,
     ): Node | null {
         if (!existingNode) {
             throw Error(PARAM_IS_EMPTY + "existingNode");
@@ -288,7 +290,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
 
     public addParentNode(
         newNodeInfo: NodeData,
-        existingNode: Node
+        existingNode: Node,
     ): Node | null {
         if (!existingNode) {
             throw Error(PARAM_IS_EMPTY + "existingNode");
@@ -417,7 +419,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
         this.selectNodeHandler.addToSelection(node);
 
         this._getNodeElementForNode(node).select(
-            mustSetFocus === undefined ? true : mustSetFocus
+            mustSetFocus === undefined ? true : mustSetFocus,
         );
 
         this.saveState();
@@ -508,7 +510,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
 
     public _triggerEvent(
         eventName: string,
-        values?: Record<string, unknown>
+        values?: Record<string, unknown>,
     ): JQuery.Event {
         const event = jQuery.Event(eventName, values);
         this.element.trigger(event);
@@ -518,18 +520,18 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
     public _openNode(
         node: Node,
         slide = true,
-        onFinished: OnFinishOpenNode | null
+        onFinished: OnFinishOpenNode | null,
     ): void {
         const doOpenNode = (
             _node: Node,
             _slide: boolean,
-            _onFinished: OnFinishOpenNode | null
+            _onFinished: OnFinishOpenNode | null,
         ): void => {
             const folderElement = new FolderElement(_node, this);
             folderElement.open(
                 _onFinished,
                 _slide,
-                this.options.animationSpeed
+                this.options.animationSpeed,
             );
         };
 
@@ -755,7 +757,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
         this.tree = new this.options.nodeClass(
             null,
             true,
-            this.options.nodeClass
+            this.options.nodeClass,
         );
 
         this.selectNodeHandler.clear();
@@ -843,7 +845,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
                 } else {
                     this.saveStateHandler.setInitialStateOnDemand(
                         state,
-                        cbFinished
+                        cbFinished,
                     );
 
                     return true;
@@ -904,7 +906,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
     }
 
     private handleClick = (
-        e: JQuery.ClickEvent<HTMLElement, any, HTMLElement, HTMLElement>
+        e: JQuery.ClickEvent<HTMLElement, any, HTMLElement, HTMLElement>,
     ): void => {
         const clickTarget = this.getClickTarget(e.target);
 
@@ -929,7 +931,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
     };
 
     private handleDblclick = (
-        e: JQuery.DoubleClickEvent<HTMLElement, any, HTMLElement, HTMLElement>
+        e: JQuery.DoubleClickEvent<HTMLElement, any, HTMLElement, HTMLElement>,
     ): void => {
         const clickTarget = this.getClickTarget(e.target);
 
@@ -981,7 +983,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
     }
 
     private handleContextmenu = (
-        e: JQuery.ContextMenuEvent<HTMLElement, any, HTMLElement, HTMLElement>
+        e: JQuery.ContextMenuEvent<HTMLElement, any, HTMLElement, HTMLElement>,
     ) => {
         const $div = jQuery(e.target).closest("ul.jqtree-tree .jqtree-element");
         if ($div.length) {
@@ -1054,7 +1056,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
 
     private doSelectNode(
         node: Node | null,
-        optionsParam?: SelectNodeOptions
+        optionsParam?: SelectNodeOptions,
     ): void {
         const saveState = (): void => {
             if (this.options.saveState) {
@@ -1157,7 +1159,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
     private doLoadDataFromUrl(
         urlInfoParam: string | JQuery.AjaxSettings | null,
         parentNode: Node | null,
-        onFinished: HandleFinishedLoading | null
+        onFinished: HandleFinishedLoading | null,
     ): void {
         const urlInfo = urlInfoParam || this.getDataUrlInfo(parentNode);
 
@@ -1167,7 +1169,7 @@ export class JqTreeWidget extends MouseWidget<JQTreeOptions> {
     private loadFolderOnDemand(
         node: Node,
         slide = true,
-        onFinished: OnFinishOpenNode | null
+        onFinished: OnFinishOpenNode | null,
     ): void {
         node.is_loading = true;
